ex: move exercice credential to the right directory
