+++ steps/auth-credential-1/app/index.js
@@ -3,7 +3,7 @@ import { setupErrorListener } from '../../../common/utils/setupErrorListener';
 import '../../../common/app/components/HomeComponent';
 import '../../../common/app/components/PeopleCardComponent';
 import '../../../common/app/components/PeopleListComponent';
-import './app/components/LoginComponent';
+import './components/LoginComponent';
 import '../../../common/app/components/PeopleApp';
 
 import 'normalize.css';
